requirements.txt
Django==5.2.4
django-bootstrap-modal-forms==3.0.5
django-bootstrap5==25.1